refactor: RandomMovableStrategy 클래스 레이어 패키지 이동
