+++ src/main/java/step3/domain/RandomMovableStrategy.java
@@ -1,4 +1,4 @@
-package step3.application;
+package step3.domain;
 
 import step3.domain.MovableStrategy;
