+++ src/test/java/step3/domain/RacingCarFactoryTest.java
@@ -2,7 +2,6 @@ package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.application.RandomMovableStrategy;
 
 import java.util.Arrays;
 import java.util.List;
